+++ catalogo_filmes/lib/components/card_movie_item_old.dart
@@ -1,4 +1,4 @@
-import 'package:catalogo_filmes/models/movie.dart';
+import 'package:catalogo_filmes/models/movie_old.dart';
 import 'package:flutter/material.dart';
 import 'package:percent_indicator/percent_indicator.dart';
 import 'package:catalogo_filmes/utils/app_routes.dart';
